requirements.txt
absl-py==2.0.0
altgraph==0.17.4
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
glcontext==2.5.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
moderngl==5.9.0
moderngl-window==2.4.5
multipledispatch==1.0.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
Pillow==10.0.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyglet==2.0.10
pyinstaller-hooks-contrib==2023.11
pyparsing==3.1.1
pyqtgraph==0.13.3
pyrr==0.10.3
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
shiboken6==6.6.1
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
